Use native stack again.
