+++ mobile-packages/app-bird-mobile/app.json
@@ -2,7 +2,7 @@
   "expo": {
     "name": "Bird",
     "slug": "app-mobile-bird",
-    "version": "1.0.2",
+    "version": "1.0.3",
     "orientation": "portrait",
     "icon": "./assets/images/birdMobileIcon.png",
     "scheme": "bird",
